updated haming for helper module and move its test to seperate file
